refactor: remove pkg dir
