+++ internal/models/checkins.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	"github.com/nathanhollows/Rapua/pkg/db"
+	"github.com/nathanhollows/Rapua/db"
 )
 
 type CheckIn struct {
