+++ internal/models/events.go
@@ -3,7 +3,7 @@ package models
 import (
 	"context"
 	"github.com/google/uuid"
-	"github.com/nathanhollows/Rapua/pkg/db"
+	"github.com/nathanhollows/Rapua/db"
 )
 
 type Event struct {
